- BUG Fix: Message-Authenticator has to be performed after serializing all attributes; - Split Radius.Packet code (lib/radius/packet.ex).
